Changed LoadCell to inherit from arduino.Arduino
